feat(iframe-api): Device handling.
