+++ react/features/device-selection/functions.js
@@ -1,5 +1,10 @@
 // @flow
-import { getAudioOutputDeviceId } from '../base/devices';
+import {
+    getAudioOutputDeviceId,
+    setAudioInputDevice,
+    setAudioOutputDeviceId,
+    setVideoInputDevice
+} from '../base/devices';
 import JitsiMeetJS from '../base/lib-jitsi-meet';
 import { toState } from '../base/redux';
 
@@ -28,4 +33,73 @@ export function getDeviceSelectionDialogProps(stateful: Object | Function) {
         selectedAudioOutputId: getAudioOutputDeviceId(),
         selectedVideoInputId: settings.cameraDeviceId
     };
+}
+
+/**
+ * Processes device requests from external applications.
+ *
+ * @param {Dispatch} dispatch - The redux {@code dispatch} function.
+ * @param {Function} getState - The redux function that gets/retrieves the redux
+ * state.
+ * @param {Object} request - The request to be processed.
+ * @param {Function} responseCallback - The callback that will send the
+ * response.
+ * @returns {boolean}
+ */
+export function processRequest(dispatch: Dispatch<*>, getState: Function, request: Object, responseCallback: Function) { // eslint-disable-line max-len, max-params
+    if (request.type === 'devices') {
+        const state = getState();
+        const settings = state['features/base/settings'];
+
+        switch (request.name) {
+        case 'isDeviceListAvailable':
+            responseCallback(JitsiMeetJS.mediaDevices.isDeviceListAvailable());
+            break;
+        case 'isDeviceChangeAvailable':
+            responseCallback(
+                JitsiMeetJS.mediaDevices.isDeviceChangeAvailable(
+                    request.deviceType));
+            break;
+        case 'isMultipleAudioInputSupported':
+            responseCallback(JitsiMeetJS.isMultipleAudioInputSupported());
+            break;
+        case 'getCurrentDevices':
+            responseCallback({
+                audioInput: settings.micDeviceId,
+                audioOutput: getAudioOutputDeviceId(),
+                videoInput: settings.cameraDeviceId
+            });
+            break;
+        case 'getAvailableDevices':
+            responseCallback(getState()['features/base/devices']);
+            break;
+        case 'setDevice': {
+            const { device } = request;
+
+            switch (device.kind) {
+            case 'audioinput':
+                dispatch(setAudioInputDevice(device.id));
+                break;
+            case 'audiooutput':
+                setAudioOutputDeviceId(device.id, dispatch);
+                break;
+            case 'videoinput':
+                dispatch(setVideoInputDevice(device.id));
+                break;
+            default:
+
+            }
+
+            responseCallback(true);
+            break;
+        }
+        default:
+
+            return false;
+        }
+
+        return true;
+    }
+
+    return false;
 }
